requirements.txt
python-telegram-bot==13.15
psycopg2==2.9.9
python-dotenv==1.0.1